added layout router, added redirection to error inside the home page
