+++ src/components/LayoutRouter/index.tsx
@@ -1,0 +1,1 @@
+export * from './LayoutRouter';
