+++ src/pages/NotFoundPage/NotFoundRedirect.ts
@@ -1,0 +1,12 @@
+import { FC, useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
+
+export const NotFoundRedirect: FC = () => {
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    navigate('/not-found');
+  }, []);
+
+  return null;
+};
